refactored models into their own packages
