+++ src/main/kotlin/org/allenterescenco/tournamentmanager/console/controllers/TournamentController.kt
@@ -2,8 +2,8 @@ package org.allenterescenco.tournamentmanager.console.controllers
 
 import mu.KotlinLogging
 
-import org.allenterescenco.tournamentmanager.console.models.TournamentMemStore
-import org.allenterescenco.tournamentmanager.console.models.TournamentModel
+import org.allenterescenco.tournamentmanager.console.models.tournament.TournamentMemStore
+import org.allenterescenco.tournamentmanager.console.models.tournament.TournamentModel
 import org.allenterescenco.tournamentmanager.console.views.TournamentView
 
 class TournamentController {
@@ -82,26 +82,32 @@ class TournamentController {
     }
 
     fun dummyData() {
-        tournaments.create(TournamentModel(name = "Pro League",
+        tournaments.create(
+            TournamentModel(name = "Pro League",
             org = "Pro Org",
             startDate = "11/06/2022",
             endDate = "15/06/2022",
             maxTeams = 64,
             partTeams = arrayOf("Team1", "Team2"),
-            winner = "Team1"))
-        tournaments.create(TournamentModel(name = "Inter League",
+            winner = "Team1")
+        )
+        tournaments.create(
+            TournamentModel(name = "Inter League",
             org = "Inter Org",
             startDate = "25/10/2022",
             endDate = "28/10/2022",
             maxTeams = 2,
             partTeams = arrayOf("Team1", "Team2"),
-            winner = ""))
-        tournaments.create(TournamentModel(name = "Rookie League",
+            winner = "")
+        )
+        tournaments.create(
+            TournamentModel(name = "Rookie League",
             org = "Rookie Org",
             startDate = "25/07/2022",
             endDate = "28/07/2022",
             maxTeams = 32,
             partTeams = arrayOf("Team1", "Team2"),
-            winner = "Team1"))
+            winner = "Team1")
+        )
     }
 }
