--- src/main/kotlin/org/allenterescenco/tournamentmanager/console/models/TeamModel.kt
@@ -1,2 +1,0 @@
-package org.allenterescenco.tournamentmanager.console.models
-
