--- src/main/kotlin/org/allenterescenco/tournamentmanager/console/models/TeamStore.kt
@@ -1,4 +1,0 @@
-package org.allenterescenco.tournamentmanager.console.models
-
-interface TeamStore {
-}
