+++ src/main/kotlin/org/allenterescenco/tournamentmanager/console/models/team/TeamModel.kt
@@ -1,0 +1,2 @@
+package org.allenterescenco.tournamentmanager.console.models.team
+
