+++ src/main/kotlin/org/allenterescenco/tournamentmanager/console/models/team/TeamStore.kt
@@ -1,0 +1,4 @@
+package org.allenterescenco.tournamentmanager.console.models.team
+
+interface TeamStore {
+}
